build: only keep production deps on final image

+++ api/entrypoint.sh
@@ -1,10 +1,28 @@
 #!/bin/sh
 
+# ====================
+# === Prisma Setup ===
+# ====================
+
+PACKAGE_JSON_PATH="./package.json"
+
+VERSION=$(jq -r ".dependencies.\"prisma\" // .devDependencies.\"prisma\"" "$PACKAGE_JSON_PATH")
+
+# Check if the version was found
+if [ -z "$VERSION" ]; then
+    echo "Error: Could not find version for prisma in $PACKAGE_JSON_PATH"
+    exit 1
+fi
+
+# Clean up the version string (remove carets, tildes, etc.)
+# This is important because dlx expects an exact version or a tag like 'latest'
+VERSION=$(echo "$VERSION" | sed 's/[^0-9.]*//g')
+
 # =========================
 # === Prisma Workaround ===
 # =========================
 
-pnpm exec prisma generate
+pnpm dlx "prisma@$VERSION" generate
 
 rm -r /app/dist/_generated/prisma
 
@@ -16,7 +34,7 @@ rm -r /app/src
 # === DB Migration ===
 # ====================
 
-pnpx prisma migrate deploy
+pnpm dlx "prisma@$VERSION" migrate deploy
 
 # =========================
 # === Start Application ===
